Move routing related functions into their own header file
