+++ src/CMakeLists.txt
@@ -74,6 +74,7 @@ set(SOURCES
     dot11/dot11_control.cpp
     utils/checksum_utils.cpp
     utils/frequency_utils.cpp
+    utils/routing_utils.cpp
 )
 
 SET(PCAP_DEPENDENT_SOURCES
